Develop

+++ src/Core/Workers/WorkerCdr.php
@@ -174,7 +174,7 @@ class WorkerCdr extends WorkerBase
     private function checkNoAnswerCall($row): void
     {
         if ($row['disposition'] === 'ANSWERED') {
-            $this->no_answered_calls[$row['linkedid']]['NOANSWER'] = false;
+            $this->no_answered_calls[$row['linkedid']]['ANSWERED'] = true;
             return;
         }
         if ( ! array_key_exists($row['dst_num'], $this->internal_numbers)) {
@@ -222,9 +222,7 @@ class WorkerCdr extends WorkerBase
         foreach ($arr_update_cdr as $data) {
             $linkedId = $data['tmp_linked_id'];
             $data['GLOBAL_STATUS'] = $data['disposition'];
-
-            $isNOANSWER = $this->no_answered_calls[$linkedId]['NOANSWER']??false;
-            if ($isNOANSWER === false) {
+            if (isset($this->no_answered_calls[$linkedId]['ANSWERED'])) {
                 $data['GLOBAL_STATUS'] = 'ANSWERED';
                 // Это отвеченный вызов (на очередь). Удаляем из списка.
                 $idForDelete[$linkedId]=true;
@@ -250,15 +248,21 @@ class WorkerCdr extends WorkerBase
             $billsec = 0;
 
             if (!empty($row['recordingfile'])) {
-                // Удаляем файлы
-                $p_info = pathinfo($row['recordingfile']);
-                $fileName = $p_info['dirname'] . '/' . $p_info['filename'];
-                $file_list = [$fileName . '.mp3', $fileName . '.wav', $fileName . '_in.wav', $fileName . '_out.wav',];
-                foreach ($file_list as $file) {
-                    if (!file_exists($file) || is_dir($file)) {
-                        continue;
+                if($row['dst_chan'] === "App:{$row['dst_num']}"){
+                    // Для приложения не может быть записи разговора.
+                    // Запись должна относится к конечному устройству.
+                    $row['recordingfile'] = '';
+                }else{
+                    // Удаляем файлы
+                    $p_info = pathinfo($row['recordingfile']);
+                    $fileName = $p_info['dirname'] . '/' . $p_info['filename'];
+                    $file_list = [$fileName . '.mp3', $fileName . '.wav', $fileName . '_in.wav', $fileName . '_out.wav',];
+                    foreach ($file_list as $file) {
+                        if (!file_exists($file) || is_dir($file)) {
+                            continue;
+                        }
+                        Processes::mwExec("rm -rf '{$file}'");
                     }
-                    Processes::mwExec("rm -rf '{$file}'");
                 }
             }
         } elseif (trim($row['recordingfile']) !== '') {
